ZeroFront -Array2-CodingBat
Return an array that contains the exact same numbers as the given array, but rearranged so that all the zeros are grouped at the start of the array. The order of the non-zero numbers does not matter. So {1, 0, 0, 10} becomes {0 ,0, 10, 1}. You may modify and return the given array or make a new array.


zeroFront([1, 0, 0, 1]) → [0, 0, 1, 1]
zeroFront([0, 1, 11, 0, 1]) → [0, 0, 1, 11, 1]
zeroFront([1, 0]) → [0, 1]