Add files via upload
PROBLEM STATEMENT
An Employee class has many attributes to it and is identified uniquely by its Id
variable. Given an Employee’s Id one should be able to retrieve the details of an
Employee.
An Employee class is given to you. The attributes of the Employee, the
getters/setters and a constructor are defined. You can add to this class if you feel
the need. But do not remove or delete existing fields/methods.
Create a class called CollectionMap. This class should maintain the mapping of
EmployeeId to the Employee object that contains its details.
Define a Map attribute called employeeMap. The mapping between a employeeId
to the employeeDetails(Employee object) are maintained in this attribute.
Define getter and setter for the map attribute. The default constructor must be
defined.
Add the following methods to the class CollectionMap:
Method: addEmployee(Employee emp):
This method should add an entry in the map for the given Employee.
● Null entries should not be allowed in the mapping.
● Return 0 for success
● Return 1 for error
● Return 2 if a duplicate entry
Method: removeEmployee(employeeId):
This removes the entry in the map for the given EmployeeId.
● Return 0 for success
● Return 1 for error
● Return 2 if the Employee is not found
Method: findEmployee(employeeId):
Given the employeeId, this method should return the corresponding Employee.
● Return the Employee object upon success
● Returns null for error or if Employee doesn’t exist.
Method: getEmployeeList():
This method returns all the employees in the map as a list.
● Returns null if the map is empty.
Test the CollectionMap class and its method with various inputs
Use the following Employee Class to define the CollectionMap Class:
public class Employee {

  int id;
  String name;
  double salary;

  public Employee() { }

  public Employee(int id, String name, double salary) {
    this.id = id;
    this.name = name;
    this.salary = salary;
 }
  public int getId() {
  return id;
 }
  public void setId(int id) {
  this.id = id;
 }
  public String getName() {
  return name;
 }
  public void setName(String name) {
  this.name = name;
 }
  public double getSalary() {
  return salary;
 }
  public void setSalary(double salary) {
  this.salary = salary;
 }
  @Override
  public String toString() {
  return "Employee [id=" + id + ", name=" + name + ", salary=" + salary+ "]";
        }
 }
Note:
Add a class called TestEmployee for testing the above Employee & CollectionMap
Classes.